Added torch.[de]serialize{To/From}Storage to File.lua.
These match the previous [de]serialize functions, but return
a torch.CharStorage instead of a Lua string. This can be used to
prevent hitting the LuaJIT 2GB limit for very large objects, because the
CharStorage is allocated on the C side (not counting towards this limit)
but the Lua string will count towards this limit.

This modification fixed a repeatable OOM crash I found for trying to
serialize certain large models.