fixed inconsistency in Tensor.__tostring__
The string that `Tensor.__tostring__` function returns includes size of different dimensions of the tensor (`tensor:size()`), but not the dimension of the tensor (`tensor:dim()`). For example, the dimension of the `torch.rand(7)`, which is obtained by `torch.rand(7):dim()` is 1, whereas Tensor.__tostring__ returns  
`[torch.DoubleTensor of dimension 7]`,
which is incorrect. This is a tensor of dimension 1, size 7. The same problem exists for Tensors of dimension 2 and beyond. For instance, for `torch.rand(8,5)` the Tensor.__tostring__ function returns
`[torch.DoubleTensor of dimension 8x5]`
instead of returning `[torch.DoubleTensor of dimension 2]` as expected by `torch.rand(8,5):dim()`.

To fix this issue and make the returned description consistent with the terminology of Torch package (e.g. Tensor:dim() and Tensor:size()), the strings `' of dimension '` have been changed to `' of size '`.